pyproject.toml
[tool.poetry]
name = "fastapi-course"
version = "0.1.0"
description = ""
authors = ["Dmitriy <ka1deo@mail.ru>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "0.105.0"
uvicorn = "^0.29.0"
alembic = "^1.13.1"
asyncpg = "^0.29.0"
python-dotenv = "^1.0.1"
dishka = "^1.0.0"
psycopg = {extras = ["binary"], version = "^3.1.18"}
psycopg2 = "^2.9.9"
flake8 = "^7.0.0"
python-jose = "^3.3.0"
bcrypt = "^4.1.2"
python-multipart = "^0.0.9"
fastapi-cache2 = {extras = ["redis"], version = "^0.2.1"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
